refactor: export mod content to dedicated package
